novena: crude first bitfile build
This is a hack job on the GPBB example bitfile with all coregen'd stuff
removed (clocks, etc).

This state of the repository should be cleaned or removed. I will next
create a template implementing the EIM and i2c interfaces (still just
copying from the GPBB example), and that might get copied in here
eventually, though I woudl like to keep the basic-hdl-template repo as
simple as possible.